Sync

+++ spec_scripts/gen_checkpoints.py
@@ -16,11 +16,11 @@ parser = argparse.ArgumentParser(
 
 parser.add_argument('--interval',
                 type=int,
-                default=10**3,
+                default=10**7,
                 help='Set simpoints interval duration in instructions')
 parser.add_argument('--warmup',
                 type=int,
-                default=10**1,
+                default=10**6,
                 help='Set simpoints warmup duration in instructions')
 
 parser.add_argument('--gem5dir',
@@ -30,17 +30,17 @@ parser.add_argument('--gem5dir',
 
 parser.add_argument('--workdir',
                 type=str,
-                default="/home/michal/Desktop/spec_2017_rate_checkpoints_gcc_10",
+                default="/home/michal/Desktop/spec_2017_rate_checkpoints",
                 help='Path to input/output directory')
 
-parser.add_argument('--specexe',
+parser.add_argument('--specdir',
                 type=str,
-                default="/home/michal/Desktop/spec_2017_rate_executables_gcc_10",
+                default="/home/michal/Desktop/spec_2017_rate_executables",
                 help='Path to minispec directory)')
 
-parser.add_argument('-n', '--nthreads',
+parser.add_argument('-j', '--jobs',
                 type=int,
-                default=2,
+                default=os.cpu_count(),
                 help='Number of jobs to run in parallel')
 
 
@@ -54,11 +54,26 @@ parser.add_argument('--clean',
                 default=False,
                 help='Clean existing ')
 
+parser.add_argument('--debug',
+                action='store_true',
+                default=False,
+                help='Use gem5 .debug build instead of .opt')
+
+parser.add_argument('--cpt-hack',
+                action='store_true',
+                default=False,
+                help='Rename all "switch_cpu_1" references in .cpt files to "cpu".')
+
 args = parser.parse_args()
 
 #Ensure absolute paths
 args.workdir = os.path.abspath(args.workdir)
-args.specexe = os.path.abspath(args.specexe)
+args.specdir = os.path.abspath(args.specdir)
+
+if(args.cpt_hack):
+    #HACK: Rename serialised structures in checkpoint to enable proper loading
+    os.system(f"""/bin/bash -c "shopt -s globstar; sed -i 's/switch_cpus_1/cpu/g' {args.workdir}/**/m5.cpt " """)
+    sys.exit()
 
 if args.clean:
     os.system(f'/bin/bash -c "shopt -s globstar; rm -rf {args.workdir}/**/checkpoints"')
@@ -72,7 +87,8 @@ if args.clean:
 #Construct list of commands to be executed in parallel
 commands = []
 sppaths = glob.glob(args.workdir + "/**/simpoints.done", recursive=True)
-benchexepaths = glob.glob(args.specexe + "/**/*.mytest-m64", recursive=True)
+sppaths.sort()
+benchexepaths = glob.glob(args.specdir + "/**/*.mytest-m64", recursive=True)
 
 #Emit command for each simpoint path
 for sppath in sppaths:
@@ -80,80 +96,98 @@ for sppath in sppaths:
     spname = sppath.split('/')[-3]
     spidx = int(sppath.split('/')[-2])
 
-    # #Skip generating clusters if issues present
-    # if os.path.exists(spdir + '/checkpoints.done'):
-    #     with open(spdir + '/checkpoints.done', 'r') as exitcode:
-    #         if int(exitcode.read().strip()) == 0:
-    #             print(f'Skipped {spdir} (Checkpoints.done with code 0)')
-    #             continue
-    
-    if os.path.exists(spdir + '/simpoints.done'):
+    #Check simpoint files
+    if os.path.exists(spdir + '/simpoints.done') and os.path.exists(spdir + '/simpoints.simpts'):
         with open(spdir + '/simpoints.done', 'r') as exitcode:
             if int(exitcode.read().strip()) != 0:
                 print(f'Skipped {spdir} (simpoint.done with non-zero exit code)')
                 continue
     else:
-        print(f'Skipped {spdir} (simpoint.done doesnt exist)')
+        print(f'Skipped {spdir} (simpoint.done/simpoints.simpts doesnt exist)')
         pass
-
-    benchexepath = [x for x in benchexepaths if spname in x][0]
-
-    benchopts = ' '.join(workloads[spname].args[spidx])
-
-    if workloads[spname].std_inputs is not None:
-        benchinfile = workloads[spname].std_inputs[spidx]
-    else:
-        benchinfile = None
-
-    benchexename = spname.split('.')[1] 
-    benchexepath = f'{args.specexe}/{spname}/{benchexename}_base.mytest-m64'
-    benchexedir = f'{args.specexe}/{spname}'
-
-    if not os.path.exists(f'{spdir}/checkpoints'):
-        os.makedirs(f'{spdir}/checkpoints')
     
-    command = []
-    command.extend([f'{args.gem5dir}/build/X86/gem5.debug', 
-                    f'--outdir={spdir}/checkpoints', 
-                    f'{args.gem5dir}/configs/example/se.py',
-                    '--cpu-type=X86KvmCPU',
-                    
-                    #My hacked SE script to switch from Kvm cpu before taking CP
-                    '--checkpoint-with-cpu=X86AtomicSimpleCPU',
-                    '--checkpoint-with-cpu-n=1',  
+    #Get number of checkpoints
+    with open(spdir + '/simpoints.simpts', 'r') as simpoints:
+        n_checkpoints = len(simpoints.readlines())
+        print(f'Found {n_checkpoints} simpoint intervals for {spdir}')
+    
+    for checkpoint_idx in range(n_checkpoints):
 
-                    f'--take-simpoint-checkpoint={spdir}/simpoints.simpts,{spdir}/simpoints.weights,{args.interval},{args.warmup}',
-                    f'--cmd={benchexepath}',
-                    f'--options={benchopts}',
-                    f'--mem-size={args.memsize}GB'])
-    
-    if benchinfile is not None:
-        command.extend(['--input', f'{args.specexe}/{spname}/{benchinfile}'])
-    
-    commands.append((spdir, benchexedir, command))
+        #Skip generating clusters if this checkpoint already generated
+        if os.path.exists(spdir + '/checkpoints.done'):
+            with open(spdir + '/checkpoints.done', 'r') as checkpoints_done:
+                done_checkpoints = checkpoints_done.read().split(' ')
+                if str(checkpoint_idx) in done_checkpoints:
+                    print(f'Skipped {spdir} CP: {checkpoint_idx} (Checkpoints.done contains checkpoint with code 0)')
+                    continue
+
+        benchexepath = [x for x in benchexepaths if spname in x][0]
+
+        benchopts = ' '.join(workloads[spname].args[spidx])
+
+        if workloads[spname].std_inputs is not None:
+            benchinfile = workloads[spname].std_inputs[spidx]
+        else:
+            benchinfile = None
+
+        benchexename = spname.split('.')[1] 
+        benchexepath = f'{args.specdir}/{spname}/{benchexename}_base.mytest-m64'
+        benchexedir = f'{args.specdir}/{spname}'
+
+        if not os.path.exists(f'{spdir}/checkpoints'):
+            os.makedirs(f'{spdir}/checkpoints')
+        
+        command = []
+        command.extend([f'{args.gem5dir}/build/X86/gem5' + ('.debug' if args.debug
+                    else '.opt'), 
+                        f'--outdir={spdir}/checkpoints', 
+                        f'{args.gem5dir}/configs/example/se.py',
+                        '--cpu-type=X86KvmCPU',
+                        
+                        #My hacked SE script to switch from Kvm cpu before taking CP
+                        '--checkpoint-with-cpu=X86AtomicSimpleCPU',
+
+                        #Simpoints 1 indexed
+                        f'--checkpoint-with-cpu-n={checkpoint_idx + 1}',  
+
+                        f'--take-simpoint-checkpoint={spdir}/simpoints.simpts,{spdir}/simpoints.weights,{args.interval},{args.warmup}',
+                        f'--cmd={benchexepath}',
+                        f'--options={benchopts}',
+                        f'--mem-size={args.memsize}GB'])
+        
+        if benchinfile is not None:
+            command.extend(['--input', f'{args.specdir}/{spname}/{benchinfile}'])
+        
+        commands.append((spdir, benchexedir, command, checkpoint_idx))
 
 #Function for single blocking program call
 def run_command(command_tuple):
-    spdir, benchexedir, command = command_tuple
+    spdir, benchexedir, command, checkpoint_idx = command_tuple
 
-    print(f"Running {spdir}")
+    print(f"Running {spdir} CP:{checkpoint_idx}")
 
-    with open(spdir + "/checkpoints.log", 'w+') as log:
+    with open(spdir  + f"/checkpoints/checkpoints{checkpoint_idx}.log", 'w+') as log:
         log.write(' '.join(command))
         process = subprocess.Popen(command, stdout=log, stderr=log, cwd=benchexedir)
         (output, err) = process.communicate()  
         p_status = process.wait()
     
+    # if os.path.exists(spdir + '/checkpoints.done'):
 
-    with open(spdir + "/checkpoints.done", 'w+') as statusf:
-        statusf.write(str(p_status))
+    if(p_status == 0):
+        with open(spdir + "/checkpoints.done", 'a+') as statusf:
+            statusf.write(str(checkpoint_idx) + " ")
     
-    print(f"Finished {spdir} with exit code {p_status}")
+    print(f"Finished {spdir} CP:{checkpoint_idx} with exit code {p_status}")
         
 #Execute commands in parallel with pool
-pool = multiprocessing.Pool(args.nthreads)
+pool = multiprocessing.Pool(args.jobs)
 
 with pool:
     pool.map(run_command, commands)
 
-print("Done")
+print("Done Running; Applying checkpoint HACK")
+
+#HACK: Rename serialised structures in checkpoint to enable proper loading
+os.system(f"""/bin/bash -c "shopt -s globstar; sed -i 's/switch_cpus_1/cpu/g' {args.workdir}/**/m5.cpt " """)
+
